some cleanup

+++ setup.py
@@ -156,24 +156,34 @@ class BuildCovalent(build_py):
     """Build Covalent with NATS server"""
 
     def run(self):
-        import subprocess
+        if not shutil.which("nats-server"):
+            import subprocess
 
-        if platform.system() == "Darwin":
-            subprocess.run(["brew", "install", "nats-server"], check=True)
-        elif platform.system() == "Linux":
-            import requests
+            if platform.system() == "Darwin":
+                subprocess.run(["brew", "install", "nats-server"], check=True)
+            elif platform.system() == "Linux":
+                import requests
 
-            r = requests.get(
-                "https://github.com/nats-io/nats-server/releases/download/v2.7.4/nats-server-v2.7.4-linux-amd64.zip",
-                allow_redirects=True,
-            )
-            open("nats-server-v2.7.4-linux-amd64.zip", "wb").write(r.content)
-            subprocess.run(["unzip", "nats-server-v2.7.4-linux-amd64.zip"], check=True)
-            shutil.move(
-                "nats-server-v2.7.4-linux-amd64/nats-server", "refactor/queuer/nats-server"
-            )
-            shutil.rmtree("nats-server-v2.7.4-linux-amd64")
-            os.remove("nats-server-v2.7.4-linux-amd64.zip")
+                r = requests.get(
+                    "https://github.com/nats-io/nats-server/releases/download/v2.7.4/nats-server-v2.7.4-linux-amd64.zip",
+                    allow_redirects=True,
+                )
+                r.raise_for_status()
+
+                open("nats-server-v2.7.4-linux-amd64.zip", "wb").write(r.content)
+                subprocess.run(["unzip", "nats-server-v2.7.4-linux-amd64.zip"], check=True)
+                shutil.move(
+                    "nats-server-v2.7.4-linux-amd64/nats-server", "refactor/queuer/nats-server"
+                )
+
+                shutil.rmtree("nats-server-v2.7.4-linux-amd64")
+                os.remove("nats-server-v2.7.4-linux-amd64.zip")
+            else:
+                print(
+                    "Platform is not natively supported. Please manually install nats-server.",
+                    file=sys.stderr,
+                )
+                sys.exit(1)
 
         build_py.run(self)
